browser updates
