+++ src/applications/chromium/sqlite.ts
@@ -1,4 +1,4 @@
-import { ChromiumProfiles, ChromiumHistory, ChromiumDownloads, ChromiumCookies, ChromiumAutofill } from "../../../types/applications/chromium";
+import { ChromiumProfiles, ChromiumHistory, ChromiumDownloads, ChromiumCookies, ChromiumAutofill, ChromiumLogins, ChromiumDips } from "../../../types/applications/chromium";
 import { PlatformType } from "../../system/systeminfo";
 import { unixEpochToISO, webkitToUnixEpoch } from "../../time/conversion";
 import { Unfold } from "../../unfold/client";
@@ -55,7 +55,7 @@ export function chromiumHistory(paths: ChromiumProfiles[], platform: PlatformTyp
                 opener_visit: entry[ "opener_visit" ] as number ?? 0,
                 unfold: undefined,
                 db_path: full_path,
-
+                version: path.version,
             };
 
             if (unfold && typeof client !== 'undefined') {
@@ -136,6 +136,7 @@ export function chromiumDownloads(paths: ChromiumProfiles[], platform: PlatformT
                 chain_index: entry[ "chain_index" ] as number ?? 0,
                 url: entry[ "url" ] as string ?? "",
                 db_path: full_path,
+                version: path.version,
             };
             hits.push(download_row);
         }
@@ -198,6 +199,7 @@ export function chromiumCookies(paths: ChromiumProfiles[], platform: PlatformTyp
                         Number(BigInt(entry[ "last_update_utc" ] as bigint) / adjust_time),
                     )),
                     db_path: full_path,
+                    version: path.version,
                 };
                 hits.push(cookie_entry);
             }
@@ -222,27 +224,221 @@ export function chromiumAutofill(paths: ChromiumProfiles[], platform: PlatformTy
             full_path = `${path.full_path}\\Default\\Web Data`;
         }
 
-        for (const path of full_path) {
-            const results = querySqlite(path, query);
-            if (results instanceof ApplicationError) {
-                console.warn(`Failed to query ${path}: ${results}`);
-                continue;
-            }
-            // Loop through cookies rows
-            const adjust_time = 1000000n;
-            for (const entry of results) {
-                const fill_entry: ChromiumAutofill = {
-                    db_path: path,
-                    date_created: unixEpochToISO(entry[ "date_created" ] as number),
-                    date_last_used: unixEpochToISO(entry[ "date_last_used" ] as number),
-                    count: entry[ "count" ] as number,
-                    name: entry[ "name" ] as string | undefined,
-                    value: entry[ "value" ] as string | undefined,
-                    value_lower: entry[ "value_lower" ] as string | undefined,
-                };
-                hits.push(fill_entry);
-            }
+        const results = querySqlite(full_path, query);
+        if (results instanceof ApplicationError) {
+            console.warn(`Failed to query ${path}: ${results}`);
+            continue;
         }
+        // Loop through cookies rows
+        const adjust_time = 1000000n;
+        for (const entry of results) {
+            const fill_entry: ChromiumAutofill = {
+                db_path: full_path,
+                date_created: unixEpochToISO(entry[ "date_created" ] as number),
+                date_last_used: unixEpochToISO(entry[ "date_last_used" ] as number),
+                count: entry[ "count" ] as number,
+                name: entry[ "name" ] as string | undefined,
+                value: entry[ "value" ] as string | undefined,
+                value_lower: entry[ "value_lower" ] as string | undefined,
+                version: path.version,
+            };
+            hits.push(fill_entry);
+        }
+
+    }
+    return hits;
+}
+
+/**
+ * Function to extract Autofill information from database
+ * @param paths Array of `ChromiumProfiles`
+ * @param platform OS `PlatformType`
+ * @param query SQL query to run
+ * @returns Array of `ChromiumLogins`
+ */
+export function chromiumLogins(paths: ChromiumProfiles[], platform: PlatformType, query: string): ChromiumLogins[] {
+    const hits: ChromiumLogins[] = [];
+    const adjust_time = 1000000n;
+
+    for (const path of paths) {
+        let full_path = `${path.full_path}/Default/Login Data`;
+
+        if (platform === PlatformType.Windows) {
+            full_path = `${path.full_path}\\Default\\Login Data`;
+        }
+
+        const data = querySqlite(full_path, query);
+        if (data instanceof ApplicationError) {
+            console.warn(`Failed to query ${path}: ${data}`);
+            continue;
+        }
+        // Loop through logins rows
+        for (const entry of data) {
+            const login_entry: ChromiumLogins = {
+                origin_url: entry[ "origin_url" ] as string,
+                signon_realm: entry[ "signon_realm" ] as string,
+                date_created: unixEpochToISO(webkitToUnixEpoch(
+                    Number(BigInt(entry[ "date_created" ] as bigint) / adjust_time),
+                )),
+                blacklisted_by_user: entry[ "blacklisted_by_user" ] as number,
+                scheme: entry[ "scheme" ] as number,
+                id: entry[ "id" ] as number,
+                date_last_used: unixEpochToISO(webkitToUnixEpoch(
+                    Number(BigInt(entry[ "date_last_used" ] as bigint) / adjust_time),
+                )),
+                date_password_modified: unixEpochToISO(webkitToUnixEpoch(
+                    Number(BigInt(entry[ "date_password_modified" ] as bigint) / adjust_time),
+                )),
+                sharing_notification_display: entry[
+                    "sharing_notification_display"
+                ] as number,
+                db_path: full_path,
+                action_url: entry[ "action_url" ] as string | undefined,
+                username_element: entry[ "username_element" ] as string | undefined,
+                username_value: entry[ "username_value" ] as string | undefined,
+                times_used: entry[ "times_used" ] as number | undefined,
+                icon_url: entry[ "icon_url" ] as string | undefined,
+                possible_username_pairs: entry[ "possible_username_pairs" ] as
+                    | string
+                    | undefined,
+                federation_url: entry[ "federation_url" ] as string | undefined,
+                generation_upload_status: entry[ "generation_upload_status" ] as
+                    | number
+                    | undefined,
+                sender_profile_image_url: entry[ "sender_profile_image_url" ] as
+                    | string
+                    | undefined,
+                password_element: entry[ "password_element" ] as string | undefined,
+                password_type: entry[ "password_type" ] as number | undefined,
+                password_value: entry[ "password_value" ] as string | undefined,
+                date_received: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "date_received" ] === "undefined"
+                        ? 0
+                        : Number(BigInt(entry[ "date_received" ] as bigint) / adjust_time),
+                )),
+                sender_email: entry[ "sender_email" ] as string | undefined,
+                sender_name: entry[ "sender_name" ] as string | undefined,
+                skip_zero_click: entry[ "skip_zero_click" ] as number | undefined,
+                submit_element: entry[ "submit_element" ] as string | undefined,
+                display_name: entry[ "display_name" ] as string | undefined,
+                form_data: entry[ "form_data" ] as string | undefined,
+                moving_blocked_for: entry[ "moving_blocked_for" ] as string | undefined,
+                keychain_identifier: entry[ "keychain_identifier" ] as string | undefined,
+                version: path.version,
+            };
+            hits.push(login_entry);
+        }
+
+    }
+    return hits;
+}
+
+/**
+ * Function to extract Detect Incidental Party State (DIPS) information from database
+ * @param paths Array of `ChromiumProfiles`
+ * @param platform OS `PlatformType`
+ * @param query SQL query to run
+ * @returns Array of `ChromiumDips`
+ */
+export function chromiumDips(paths: ChromiumProfiles[], platform: PlatformType, query: string): ChromiumDips[] {
+    const hits: ChromiumDips[] = [];
+    const adjust_time = 1000000n;
+
+    for (const path of paths) {
+        let full_path = `${path.full_path}/Default/DIPS`;
+
+        if (platform === PlatformType.Windows) {
+            full_path = `${path.full_path}\\Default\\DIPS`;
+        }
+
+        const data = querySqlite(full_path, query);
+        if (data instanceof ApplicationError) {
+            console.warn(`Failed to query ${path}: ${data}`);
+            continue;
+        }
+        // Loop through logins rows
+        for (const entry of data) {
+            const dips_entry: ChromiumDips = {
+                site: entry[ "site" ] as string,
+                path: full_path,
+                first_bounce: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "first_bounce_time" ] === "undefined" ||
+                        entry[ "first_bounce_time" ] === null
+                        ? 0
+                        : Number(BigInt(entry[ "first_bounce_time" ] as bigint) / adjust_time),
+                )),
+                last_bounce: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "last_bounce_time" ] === "undefined" ||
+                        entry[ "last_bounce_time" ] === null
+                        ? 0
+                        : Number(BigInt(entry[ "last_bounce_time" ] as bigint) / adjust_time),
+                )),
+                first_site_storage: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "first_site_storage_time" ] === "undefined" ||
+                        entry[ "first_site_storage_time" ] === null
+                        ? 0
+                        : Number(
+                            BigInt(entry[ "first_site_storage_time" ] as bigint) / adjust_time,
+                        ),
+                )),
+                first_stateful_bounce: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "first_stateful_bounce_time" ] === "undefined" ||
+                        entry[ "first_stateful_bounce_time" ] === null
+                        ? 0
+                        : Number(
+                            BigInt(entry[ "first_stateful_bounce_time" ] as bigint) / adjust_time,
+                        ),
+                )),
+                first_user_interaction: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "first_user_interaction_time" ] === "undefined" ||
+                        entry[ "first_user_interaction_time" ] === null
+                        ? 0
+                        : Number(
+                            BigInt(entry[ "first_user_interaction_time" ] as bigint) /
+                            adjust_time,
+                        ),
+                )),
+                first_web_authn_assertion: unixEpochToISO(webkitToUnixEpoch(
+                    entry[ "first_web_authn_assertion_time" ] === null ? 0 : Number(
+                        BigInt(entry[ "first_web_authn_assertion_time" ] as bigint) /
+                        adjust_time,
+                    ),
+                )),
+                last_site_storage: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "last_site_storage_time" ] === "undefined" ||
+                        entry[ "last_site_storage_time" ] === null
+                        ? 0
+                        : Number(
+                            BigInt(entry[ "last_site_storage_time" ] as bigint) / adjust_time,
+                        ),
+                )),
+                last_stateful_bounce: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "last_stateful_bounce_time" ] === "undefined" ||
+                        entry[ "last_stateful_bounce_time" ] === null
+                        ? 0
+                        : Number(
+                            BigInt(entry[ "last_stateful_bounce_time" ] as bigint) / adjust_time,
+                        ),
+                )),
+                last_user_interaction: unixEpochToISO(webkitToUnixEpoch(
+                    typeof entry[ "last_user_interaction_time" ] === "undefined" ||
+                        entry[ "last_user_interaction_time" ] === null
+                        ? 0
+                        : Number(
+                            BigInt(entry[ "last_user_interaction_time" ] as bigint) / adjust_time,
+                        ),
+                )),
+                last_web_authn_assertion: unixEpochToISO(webkitToUnixEpoch(
+                    entry[ "last_web_authn_assertion_time" ] === null ? 0 : Number(
+                        BigInt(entry[ "last_web_authn_assertion_time" ] as bigint) /
+                        adjust_time,
+                    ),
+                )),
+                version: path.version,
+            };
+            hits.push(dips_entry);
+        }
+
     }
     return hits;
 }
